+++ Network/Models/NetworkStats.cs
@@ -3,8 +3,8 @@
 public class NetworkStats
 {
     public int ConnectedClients { get; set; }
-    public int PacketsSent { get; set; }
-    public int PacketsReceived { get; set; }
+    public long PacketsSent { get; set; }
+    public long PacketsReceived { get; set; }
     public bool IsRunning { get; set; }
     public int Port { get; set; }
 }
